requirements.txt
fastapi
uvicorn
python-nmap
pydantic
boto3
pdfkit
